+++ src/modules/messenger/Messenger.ts
@@ -45,15 +45,7 @@ export default class Messenger extends Block {
 
   private _createChatForm: Nullable<HTMLFormElement>;
 
-  private _addUserForm: Nullable<HTMLFormElement>;
-
-  private _deleteUserForm: Nullable<HTMLFormElement>;
-
   private _createChatValidator: Nullable<Validator>;
-
-  private _addUserValidator: Nullable<Validator>;
-
-  private _deleteUserValidator: Nullable<Validator>;
 
   private _handlers: (() => void)[];
 
@@ -61,7 +53,6 @@ export default class Messenger extends Block {
   // eslint-disable-next-line no-useless-constructor
   constructor(props: MessengerProps, ...rest: [Children?, BlockParams?]) {
     super(props, ...rest);
-    console.log(this);
   }
 
   init(): void {
@@ -75,11 +66,7 @@ export default class Messenger extends Block {
       '.js-messenger__sidebar-close'
     ) as HTMLButtonElement;
     this._createChatForm = document.querySelector('.js-create-chat') as HTMLFormElement;
-    this._addUserForm = document.querySelector('.js-add-user') as HTMLFormElement;
-    this._deleteUserForm = document.querySelector('.js-delete-user') as HTMLFormElement;
     this._createChatValidator = new Validator(this._createChatForm);
-    this._addUserValidator = new Validator(this._addUserForm);
-    this._deleteUserValidator = new Validator(this._deleteUserForm);
 
     this._handlers = [
       addEventListener(this.element, 'store-update', (e: CustomEvent) => {
@@ -113,51 +100,6 @@ export default class Messenger extends Block {
         }
       }),
       addEventListener(this._closeButton, 'click', () => this.toggleSidebar(false)),
-      addEventListener(this.element, 'submit', async (e) => {
-        const form = e.target as HTMLFormElement;
-        const isAddUser = form === this._addUserForm;
-        const isDeleteUser = form === this._deleteUserForm;
-
-        if (!isAddUser || !isDeleteUser) {
-          return;
-        }
-
-        e.preventDefault();
-
-        if (
-          !(this._addUserValidator as Validator).valid ||
-          !(this._deleteUserValidator as Validator).valid
-        ) {
-          return;
-        }
-
-        const data = new FormData(form);
-        const spinElem = (form.querySelector('button[type="submit"]') || form) as HTMLElement;
-        let action;
-
-        if (isAddUser) {
-          action = 'add';
-        }
-
-        if (isDeleteUser) {
-          action = 'delete';
-        }
-
-        const wasOk = await this.controller.manageUser(
-          {
-            user: {
-              login: data.get(`${action}-user`) as string,
-            },
-            chatId: Number(form.dataset.chat),
-          },
-          action as 'add' | 'delete',
-          spinElem
-        );
-
-        if (wasOk) {
-          (this.getComponent(`${action}-user-modal`) as Modal).close();
-        }
-      }),
       addEventListener(this._createChatForm, 'submit', async (e) => {
         e.preventDefault();
 
@@ -172,28 +114,6 @@ export default class Messenger extends Block {
 
         if (isChatCreated) {
           (this.getComponent('chat-create-modal') as Modal).close();
-        }
-      }),
-      addEventListener(this.element, 'submit', async (e) => {
-        const form = e.target as HTMLFormElement;
-        const isDeleteChatForm = form.matches('.js-delete-chat');
-
-        if (!isDeleteChatForm) {
-          return;
-        }
-
-        e.preventDefault();
-
-        const spinElem = (form.querySelector('button[type="submit"]') || form) as HTMLElement;
-        const isChatDeleted = await this.controller.deleteChat(
-          {
-            chatId: Number(form.dataset.chat),
-          },
-          spinElem
-        );
-
-        if (isChatDeleted) {
-          (this.getComponent('delete-chat-modal') as Modal).close();
         }
       }),
     ];
@@ -218,10 +138,6 @@ export default class Messenger extends Block {
     this._chatListContainer = null;
     this._chatContainer = null;
     this._createChatForm = null;
-    this._addUserForm = null;
-    this._deleteUserForm = null;
     this._createChatValidator = null;
-    this._addUserValidator = null;
-    this._deleteUserValidator = null;
   }
 }
