+++ src/pages/login/Login.ts
@@ -1,4 +1,4 @@
-import Router from '../../utils/Router';
+import Router from '../../utils/router/Router';
 import Block from '../../modules/block';
 import Page from '../../layout/page';
 import Auth from '../../modules/auth';
