+++ src/pages/page-404/Page404.ts
@@ -1,4 +1,4 @@
-import Router from '../../utils/Router';
+import Router from '../../utils/router/Router';
 import Block from '../../modules/block';
 import Page from '../../layout/page';
 import NetworkError from '../../components/network-error';
