+++ src/store/actions.ts
@@ -1,5 +1,3 @@
-// USER_ADD
-// USER_DELETE
 /* eslint-disable no-param-reassign */
 import sortById from '../utils/sortById';
 import cloneDeep from '../utils/cloneDeep';
@@ -112,20 +110,32 @@ export default {
       previousChat.selected = false;
     }
 
-    Object.keys(state.messages as Record<string, MessageData>).forEach((key) => {
-      (state.messages as Record<string, Nullable<MessageData[]>>)[key] = null;
-    });
+    if (state.messages) {
+      Object.keys(state.messages).forEach((key) => {
+        (state.messages as Record<string, Nullable<MessageData[]>>)[key] = null;
+      });
+    }
+
     state.currentChat = null;
   },
 
   MESSAGES_LOAD(payload: { chatId: number; data: MessageData[] }, state: StateTree): void {
     const { chatId, data } = payload;
+    const { currentChat } = state;
 
     if (!state.messages) {
       state.messages = {};
     }
 
     (state.messages as Record<string, MessageData[]>)[chatId] = sortById(data) as MessageData[];
+
+    if (currentChat) {
+      const affectedChat = (state.chats as ChatData[]).find(
+        (chat) => chat.id === (currentChat as ChatData).id
+      ) as ChatData;
+
+      affectedChat.unread_count = 0;
+    }
   },
   MESSAGE_ADD(payload: { chatId: number; data: MessageData }, state: StateTree): void {
     const { chatId, data } = payload;
@@ -134,15 +144,17 @@ export default {
     const affectedChatClone = cloneDeep(affectedChat) as ChatData;
     const { last_message: lastMessage, unread_count: unreadCount } = affectedChatClone;
 
-    (state.messages as Record<string, MessageData[]>)[chatId].push(data);
-    Object.assign(lastMessage, {
+    if (state.messages && (state.messages as Record<string, MessageData[]>)[chatId]) {
+      (state.messages as Record<string, MessageData[]>)[chatId].push(data);
+    }
+
+    affectedChat.last_message = Object.assign(lastMessage || {}, {
       content: data.content,
       time: data.time,
       user: data.user_id, // извините
     });
-    affectedChat.last_message = lastMessage;
 
-    if ((currentChat as ChatData).id !== chatId) {
+    if (!currentChat || (currentChat as ChatData).id !== chatId) {
       affectedChat.unread_count = unreadCount + 1;
     }
   },
